extract MemoryWatchViewModel from MemoryBoomarkViewModel
