+++ src/RA_Integration.vcxproj.filters
@@ -432,6 +432,9 @@
     <ClCompile Include="data\search\MemBlock.cpp">
       <Filter>Data\Search</Filter>
     </ClCompile>
+    <ClCompile Include="ui\viewmodels\MemoryWatchViewModel.cpp">
+      <Filter>UI\ViewModels</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="RA_Resource.h">
@@ -1075,6 +1078,9 @@
     </ClInclude>
     <ClInclude Include="data\search\MemBlock.hh">
       <Filter>Data\Search</Filter>
+    </ClInclude>
+    <ClInclude Include="ui\viewmodels\MemoryWatchViewModel.hh">
+      <Filter>UI\ViewModels</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
