+++ src/ui/viewmodels/MemoryBookmarksViewModel.hh
@@ -10,8 +10,7 @@
 #include "services\TextWriter.hh"
 
 #include "ui\WindowViewModelBase.hh"
-
-#include "ui\viewmodels\LookupItemViewModel.hh"
+#include "ui\viewmodels\MemoryWatchViewModel.hh"
 
 namespace ra {
 namespace ui {
@@ -43,183 +42,9 @@ public:
         PauseOnChange,
     };
 
-    class MemoryBookmarkViewModel : public LookupItemViewModel,
-        protected ra::data::context::EmulatorContext::DispatchesReadMemory
+    class MemoryBookmarkViewModel : public MemoryWatchViewModel
     {
     public:
-        /// <summary>
-        /// The <see cref="ModelProperty" /> for bookmark summary.
-        /// </summary>
-        static const StringModelProperty DescriptionProperty;
-
-        /// <summary>
-        /// Gets the bookmark summary.
-        /// </summary>
-        const std::wstring& GetDescription() const { return GetValue(DescriptionProperty); }
-
-        /// <summary>
-        /// Sets the bookmark summary.
-        /// </summary>
-        void SetDescription(const std::wstring& sValue) { SetValue(DescriptionProperty, sValue); }
-
-        /// <summary>
-        /// The <see cref="ModelProperty" /> for the whether the bookmark description is custom.
-        /// </summary>
-        static const BoolModelProperty IsCustomDescriptionProperty;
-
-        /// <summary>
-        /// Gets whether the bookmark description is custom.
-        /// </summary>
-        bool IsCustomDescription() const { return GetValue(IsCustomDescriptionProperty); }
-
-        /// <summary>
-        /// Sets whether the bookmark description is custom.
-        /// </summary>
-        void SetIsCustomDescription(bool bValue) { SetValue(IsCustomDescriptionProperty, bValue); }
-
-        /// <summary>
-        /// The <see cref="ModelProperty" /> for the note supporting the bookmark.
-        /// </summary>
-        static const StringModelProperty RealNoteProperty;
-
-        /// <summary>
-        /// Gets the note supporting the bookmark.
-        /// </summary>
-        const std::wstring& GetRealNote() const { return GetValue(RealNoteProperty); }
-
-        /// <summary>
-        /// Sets the note supporting the bookmark.
-        /// </summary>
-        void SetRealNote(const std::wstring& sValue) { SetValue(RealNoteProperty, sValue); }
-
-        /// <summary>
-        /// Sets the <see cref="RealNoteProperty" /> based on the <see cref="CurrentAddressProperty" />.
-        /// </summary>
-        void UpdateRealNote();
-
-        /// <summary>
-        /// The <see cref="ModelProperty" /> for the bookmark address.
-        /// </summary>
-        static const IntModelProperty AddressProperty;
-
-        /// <summary>
-        /// Gets the bookmark address.
-        /// </summary>
-        ByteAddress GetAddress() const noexcept { return m_nAddress; }
-
-        /// <summary>
-        /// Sets the bookmark address.
-        /// </summary>
-        void SetAddress(ByteAddress value) { SetValue(AddressProperty, value); }
-
-        /// <summary>
-        /// Sets an indirect bookmark address.
-        /// </summary>
-        void SetIndirectAddress(const std::string& sSerialized);
-
-        /// <summary>
-        /// Gets the indirect bookmark address.
-        /// </summary>
-        /// <returns></returns>
-        const std::string& GetIndirectAddress() const noexcept { return m_sIndirectAddress; }
-
-        /// <summary>
-        /// Gets whether or not the bookmark has an indirect bookmark address.
-        /// </summary>
-        bool HasIndirectAddress() const noexcept { return !m_sIndirectAddress.empty(); }
-
-        /// <summary>
-        /// Updates the current address from the indirect address chain.
-        /// </summary>
-        /// <returns><c>true</c> if the indirect address chain was valid, <c>false</c> if not.</returns>
-        /// <remarks>Address will be updated to whatever the chain evaluates to, regardless of the chain's validity</remarks>
-        bool UpdateCurrentAddress();
-
-        /// <summary>
-        /// The <see cref="ModelProperty" /> for the bookmark size.
-        /// </summary>
-        static const IntModelProperty SizeProperty;
-
-        /// <summary>
-        /// Gets the bookmark size.
-        /// </summary>
-        MemSize GetSize() const noexcept { return m_nSize; }
-
-        /// <summary>
-        /// Sets the bookmark size.
-        /// </summary>
-        void SetSize(MemSize value) { SetValue(SizeProperty, ra::etoi(value)); }
-
-        /// <summary>
-        /// The <see cref="ModelProperty" /> for the bookmark format.
-        /// </summary>
-        static const IntModelProperty FormatProperty;
-
-        /// <summary>
-        /// Gets the bookmark format.
-        /// </summary>
-        MemFormat GetFormat() const { return ra::itoe<MemFormat>(GetValue(FormatProperty)); }
-
-        /// <summary>
-        /// Sets the bookmark format.
-        /// </summary>
-        void SetFormat(MemFormat value) { SetValue(FormatProperty, ra::etoi(value)); }
-
-        /// <summary>
-        /// The <see cref="ModelProperty" /> for the current value of the bookmarked address.
-        /// </summary>
-        static const StringModelProperty CurrentValueProperty;
-
-        /// <summary>
-        /// Gets the current value of the bookmarked address. 
-        /// </summary>
-        const std::wstring& GetCurrentValue() const { return GetValue(CurrentValueProperty); }
-
-        /// <summary>
-        /// Sets the current value of the bookmarked address.
-        /// </summary>
-        bool SetCurrentValue(const std::wstring& sValue, _Out_ std::wstring& sError);
-
-        /// <summary>
-        /// Gets the unformatted current value of the bookmarked address.
-        /// </summary>
-        uint32_t GetCurrentValueRaw() const noexcept { return m_nValue; }
-
-        /// <summary>
-        /// Sets the current value of the bookmarked address.
-        /// </summary>
-        void SetCurrentValueRaw(unsigned nValue);
-
-        /// <summary>
-        /// The <see cref="ModelProperty" /> for the previous value of the bookmarked address.
-        /// </summary>
-        static const StringModelProperty PreviousValueProperty;
-
-        /// <summary>
-        /// Gets the previous value of the bookmarked address.
-        /// </summary>
-        const std::wstring& GetPreviousValue() const { return GetValue(PreviousValueProperty); }
-
-        /// <summary>
-        /// Sets the previous value of the bookmarked address.
-        /// </summary>
-        void SetPreviousValue(const std::wstring& sValue) { SetValue(PreviousValueProperty, sValue); }
-
-        /// <summary>
-        /// The <see cref="ModelProperty" /> for the number of times the value of the bookmarked address changed.
-        /// </summary>
-        static const IntModelProperty ChangesProperty;
-
-        /// <summary>
-        /// Gets the number of times the value of the bookmarked address changed.
-        /// </summary>
-        unsigned int GetChanges() const { return gsl::narrow_cast<unsigned int>(GetValue(ChangesProperty)); }
-
-        /// <summary>
-        /// Sets the number of times the value of the bookmarked address changed.
-        /// </summary>
-        void SetChanges(unsigned int value) { SetValue(ChangesProperty, value); }
-
         /// <summary>
         /// The <see cref="ModelProperty" /> for the behavior of the bookmark.
         /// </summary>
@@ -236,68 +61,23 @@ public:
         void SetBehavior(BookmarkBehavior value) { SetValue(BehaviorProperty, ra::etoi(value)); }
 
         /// <summary>
-        /// The <see cref="ModelProperty" /> for whether or not the <see cref="CurrentValueProperty"/> is read only.
+        /// Reads the watched memory and updates the <see cref="CurrentValueProperty" />.
         /// </summary>
-        static const BoolModelProperty ReadOnlyProperty;
+        /// <remarks>
+        /// Should only be called in response to the watched memory being changed by the user.
+        /// Bypasses behavioral triggers.
+        /// </remarks>
+        void UpdateCurrentValue();
 
-        /// <summary>
-        /// Gets whether or not the <see cref="CurrentValueProperty"/> is read only. 
-        /// </summary>
-        bool IsReadOnly() const { return GetValue(ReadOnlyProperty); }
-
-        /// <summary>
-        /// Sets whether or not the <see cref="CurrentValueProperty"/> is read only.
-        /// </summary>
-        void SetReadOnly(bool bValue) { SetValue(ReadOnlyProperty, bValue); }
-
-        /// <summary>
-        /// The <see cref="ModelProperty" /> for the row color.
-        /// </summary>
-        static const IntModelProperty RowColorProperty;
-
-        /// <summary>
-        /// Gets the row color.
-        /// </summary>
-        Color GetRowColor() const { return Color(ra::to_unsigned(GetValue(RowColorProperty))); }
-
-        /// <summary>
-        /// Sets the row color.
-        /// </summary>
-        void SetRowColor(Color value) { SetValue(RowColorProperty, ra::to_signed(value.ARGB)); }
-
-        /// <summary>
-        /// Gets whether or not the bookmark has been modified.
-        /// </summary>
-        bool IsModified() const noexcept { return m_bModified; }
-
-        /// <summary>
-        /// Clears the modified flag.
-        /// </summary>
-        void ResetModified() noexcept { m_bModified = false; }
-
-        /// <summary>
-        /// Determines if the bookmarked memory has changed since the last time MemoryChanged was called.
-        /// </summary>
-        /// <returns><c>true</c> if the memory has changed, <c>false</c> if not.</returns>
-        bool MemoryChanged();
-
-        /// <summary>
-        /// Starts initialization of the bookmark.
-        /// </summary>
-        void BeginInitialization() noexcept { m_bInitialized = false; }
-
-        /// <summary>
-        /// Finishes initialization of the bookmark.
-        /// </summary>
-        void EndInitialization();
-
+        // ==== used to communicate user-driven change to watched memory through bookmark view model
+        // ==== TODO: add an m_bUpdatingCurrentValue flag and have UpdateCurrentValue abort if that's set
         /// <summary>
         /// The <see cref="ModelProperty" /> for whether or not the <see cref="CurrentValueProperty"/> is dirty.
         /// </summary>
         static const BoolModelProperty IsDirtyProperty;
 
         /// <summary>
-        /// Gets whether or not the <see cref="CurrentValueProperty"/> is dirty. 
+        /// Gets whether or not the <see cref="CurrentValueProperty"/> is dirty.
         /// </summary>
         bool IsDirty() const { return GetValue(IsDirtyProperty); }
 
@@ -305,38 +85,13 @@ public:
         /// Sets whether or not the <see cref="CurrentValueProperty"/> is dirty.
         /// </summary>
         void SetDirty(bool bValue) { SetValue(IsDirtyProperty, bValue); }
-
-        /// <summary>
-        /// Updates the <see cref="CurrentValueProperty" /> for the bookmark.
-        /// </summary>
-        void UpdateCurrentValue();
+        // ===== END TODO
 
     protected:
         void OnValueChanged(const IntModelProperty::ChangeArgs& args) override;
-        void OnValueChanged(const StringModelProperty::ChangeArgs& args) override;
 
-        void OnValueChanged();
-        void OnSizeChanged();
-
-        unsigned ReadValue() const;
-
-        void SetAddressWithoutUpdatingValue(ra::ByteAddress nNewAddress);
-
-    private:
-        std::wstring BuildCurrentValue() const;
-        static std::wstring ExtractDescriptionHeader(const std::wstring& sFullNote);
-
-        // keep address/size/value fields directly accessible for speed - also keep in ValueProperty for binding
-        ra::ByteAddress m_nAddress = 0;
-        uint32_t m_nValue = 0;
-        MemSize m_nSize = MemSize::EightBit;
-        bool m_bModified = false;
-        bool m_bInitialized = false;
-        bool m_bSyncingDescriptionHeader = false;
-
-        std::string m_sIndirectAddress;
-        std::unique_ptr<uint8_t[]> m_pBuffer;
-        rc_value_t* m_pValue = nullptr;
+        bool IgnoreValueChange(uint32_t nNewValue) override;
+        bool ChangeValue(uint32_t nNewValue) override;
     };
 
     /// <summary>
